% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epubr.R
\name{epub_combine}
\alias{epub_combine}
\title{Read and combine EPUB e-books}
\usage{
epub_combine(files, add_pattern = NULL, hgr_clean = FALSE,
  section_drop = NULL, series = FALSE, parent_dir = "novels")
}
\arguments{
\item{files}{character, input EPUB filename, may be a vector.}

\item{add_pattern}{supplemental regular expression pattern string to use for identifying chapters among all document sections. May be a function like, e.g. \code{\link{pat_startrek}} with additional filtering information for specific book titles.}

\item{hgr_clean}{logical, if \code{TRUE} and \code{hgr} package is installed from GitHub, an extra cleaning step is performed on the e-book text.}

\item{section_drop}{supplemental regular expression pattern string to use for dropping rows, typically non-chapter book sections. May be a function that returns a string like, e.g. \code{\link{sec_drop_startrek}}.}

\item{series}{logical, attempt to guess series and subseries from directory structure. See details.}

\item{parent_dir}{character, parent directory under which any series and subseries, if applicable, are nested. Set this whenever \code{series = TRUE}.}
}
\value{
a data frame
}
\description{
Read and combine in a data frame multiple EPUB format e-books.
}
\details{
This function will attempt some additional processing of the e-books' metadata and text content. For example, it can attempt to guess series and subseries based on the lowest two directories if files are stored in a nested directory tree.
If this does not have any meaning for the \code{files}, leave \code{series = FALSE}.
}
\examples{
# not run
}
\seealso{
\code{\link{epub_read}}
}
