% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epubr.R
\name{epub_meta}
\alias{epub_meta}
\title{Read EPUB file metadata}
\usage{
epub_meta(files, add_pattern = NULL)
}
\arguments{
\item{files}{character vector of extracted files from EPUB file archive. See details.}

\item{add_pattern}{supplemental regular expression pattern string to use for identifying chapters among all document sections. May be a function like, e.g. \code{\link{pat_startrek}} with additional filtering information for specific book titles.}
}
\value{
a data frame.
}
\description{
Parse EPUB file metadata structure into a data frame. Assumptions are made, liberties taken, some fields may be dropped. You can always inspect the complete archive contents directly after using \code{epub_unzip}.
}
\details{
This function is not typically called directly. If called directly, it must be done with \code{files} being a vector of the files created by the extraction performed by \code{epub_unzip}.
Typically this function is called internally as part of \code{epub_read}, which combines the e-book metadata with the e-book text in a single data frame.
}
\examples{
# not run
}
\seealso{
\code{\link{epub_unzip}}, \code{\link{epub_read}}
}
