% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epubr.R
\name{epub_read}
\alias{epub_read}
\title{Read EPUB file}
\usage{
epub_read(file, add_pattern = NULL, hgr_clean = FALSE)
}
\arguments{
\item{file}{character, input EPUB filename.}

\item{add_pattern}{supplemental regular expression pattern string to use for identifying chapters among all document sections. May be a function like, e.g. \code{\link{pat_startrek}} with additional filtering information for specific book titles.}

\item{hgr_clean}{logical, if \code{TRUE} and \code{hgr} package is installed from GitHub, an extra cleaning step is performed on the e-book text.}
}
\value{
a data frame
}
\description{
Parse EPUB file metadata and textual content into a data frame.
}
\details{
The data frame has metadata in all columns except the \code{data} column, which is a nested column containing a data frame of the e-book text. Both data frames are tibbles and safe to print to the console "as is".
}
\examples{
# not run
}
\seealso{
\code{\link{epub_unzip}}, \code{\link{epub_meta}}, \code{\link{epub_combine}}
}
