% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epubr.R
\name{epub_unzip}
\alias{epub_unzip}
\title{Open EPUB file archive}
\usage{
epub_unzip(file, exdir = file.path(tempdir(), basename(file)))
}
\arguments{
\item{file}{character, input EPUB file.}

\item{exdir}{character, extraction directory to place archive contents (files).}
}
\value{
nothing returned by files are extracted from archive.
}
\description{
Unzip an EPUB file.
}
\examples{
# not run
}
\seealso{
\code{\link{epub_meta}}, \code{\link{epub_read}}
}
