% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pat_startrek}
\alias{pat_startrek}
\title{Star Trek novel overrides}
\usage{
pat_startrek()
}
\value{
a named list.
}
\description{
Pattern helper function specific to Star Trek novels. A function that takes no arguments and returns a named list of three elements: \code{pattern} is a regular expression pattern string,
\code{chapter_check} is a character vector giving exact (in-metadata) book titles as a filter to only apply the supplemental \code{pattern} to specific books,
and \code{chapter_doublecheck} is a vector of titles that applies additional checks that may only be necessary and helpful (and non-harmful) for specific titles.
}
\examples{
pat_startrek()
}
