% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sec_drop_startrek}
\alias{sec_drop_startrek}
\title{Star Trek novel section filter}
\usage{
sec_drop_startrek()
}
\value{
a character string.
}
\description{
Regular expression pattern for dropping commonly named non-chapter sections that typically apply to popular Star Trek novels.
}
\examples{
sec_drop_startrek()
}
