% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{st_epub_test}
\alias{st_epub_test}
\title{Star Trek EPUB tester}
\usage{
st_epub_test(file, details = FALSE, add_tail = FALSE)
}
\arguments{
\item{file}{character, input EPUB file.}

\item{details}{logical, print more details to console.}

\item{add_tail}{logical, also print tail of data frame.}
}
\value{
nothing is returned but information is logged at the console.
}
\description{
Example testing function for parsing Star Trek novels.
}
\examples{
# not run
}
