% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{st_fix_date}
\alias{st_fix_date}
\title{Fix Star Trek date column}
\usage{
st_fix_date(x)
}
\arguments{
\item{x}{a data frame returned by a function such as \code{epub_read} or \code{epub_combine}.}
}
\value{
a data frame
}
\description{
Improve \code{date} column in EPUB data frame output for Star Trek novels by replacing their year dates with \code{yyyy-mm-dd} format date strings taken from any available (most) \code{source} column input filenames.
}
\examples{
# not run
}
