% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epubr.R
\docType{package}
\name{epubr}
\alias{epubr}
\alias{epubr-package}
\title{epubr: Read EPUB file metadata and text.}
\description{
epubr provides functions supporting the reading and parsing of internal e-book content from EPUB files. Book metadata and textual contents are parse separately.
}
\details{
E-book formatting is non-standard enough across all literature that no function can curate parsed e-book content for an across an arbitrary collection of e-books, in completely general form, resulting in a singular, consistently formatted output. Functions in this package that are intended for relatively general application to arbitrary e-books with *a priori* unknown formatting and metadata structure, do minimal to no processing of the output. They essentially only read the content "as is" and attempt to curate them in a data frame. Some poorly formatted e-books may not be readable by these functions at all.
}
